Refactoring for better inlining.  Replaced binomial heap with lazy aux pairing heap. Q14 is currently broken (again) and needs investigating.

+++ HRDBMS/src/com/exascale/optimizer/AggregateOperator.java
@@ -26,7 +26,7 @@ public interface AggregateOperator
 
 	public void setInputColumn(String col);
 
-	public void setNumGroups(int groups);
+	public void setNumGroups(long groups);
 
 	public abstract class AggregateResultThread extends ThreadPoolThread
 	{
