+++ HRDBMS/src/com/exascale/optimizer/IntersectOperator.java
@@ -54,7 +54,7 @@ public final class IntersectOperator implements Operator, Serializable
 	private int node;
 	private transient ArrayList<HashSet<ArrayList<Object>>> sets;
 	private transient BufferedLinkedBlockingQueue buffer;
-	private int estimate = 16;
+	private long estimate = 16;
 	private transient boolean inMem;
 	private transient int numFiles;
 
@@ -85,7 +85,7 @@ public final class IntersectOperator implements Operator, Serializable
 		value.cols2Pos = OperatorUtils.deserializeStringIntHM(in, prev);
 		value.pos2Col = OperatorUtils.deserializeTM(in, prev);
 		value.node = OperatorUtils.readInt(in);
-		value.estimate = OperatorUtils.readInt(in);
+		value.estimate = OperatorUtils.readLong(in);
 		value.startDone = OperatorUtils.readBool(in);
 		value.estimateSet = OperatorUtils.readBool(in);
 		value.received = new AtomicLong(0);
@@ -310,7 +310,7 @@ public final class IntersectOperator implements Operator, Serializable
 		OperatorUtils.serializeStringIntHM(cols2Pos, out, prev);
 		OperatorUtils.serializeTM(pos2Col, out, prev);
 		OperatorUtils.writeInt(node, out);
-		OperatorUtils.writeInt(estimate, out);
+		OperatorUtils.writeLong(estimate, out);
 		OperatorUtils.writeBool(startDone, out);
 		OperatorUtils.writeBool(estimateSet, out);
 	}
@@ -320,7 +320,7 @@ public final class IntersectOperator implements Operator, Serializable
 	{
 	}
 
-	public boolean setEstimate(int estimate)
+	public boolean setEstimate(long estimate)
 	{
 		if (estimateSet)
 		{
